wap-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
# Azure AI Services
openai==1.12.0
azure-search-documents==11.4.0
azure-identity==1.15.0
python-dotenv==1.0.0
redis==5.0.1
resend==0.7.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Linting and formatting
ruff==0.1.6
black==23.11.0
mypy==1.7.1

